+++ doc/CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 HEAD:
+- Added support for opaque types (newtype declarations).
 
 Version 0.8.1: